% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukhousing-package.R
\docType{package}
\name{ukhousing-package}
\alias{ukhousing}
\alias{ukhousing-package}
\title{ukhousing: Access UK Housing Data from Land Registry, EPC, and Planning}
\description{
Fetch UK housing data from official sources. Access the UK House Price Index and Price Paid Data from 'HM Land Registry' \url{https://landregistry.data.gov.uk/}, domestic and non-domestic Energy Performance Certificates from the 'MHCLG' Open Data service \url{https://epc.opendatacommunities.org/}, and planning application, brownfield land, and local plan data from 'planning.data.gov.uk' \url{https://www.planning.data.gov.uk/}. Data covers all 441 UK local authorities from 1995 to the present. Functions accept flexible filters (postcode, local authority, property type, date range) and return tidy data frames. Downloaded data is cached locally for subsequent calls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/charlescoverdale/ukhousing}
  \item Report bugs at \url{https://github.com/charlescoverdale/ukhousing/issues}
}

}
\author{
\strong{Maintainer}: Charles Coverdale \email{charlesfcoverdale@gmail.com}

}
\keyword{internal}
